% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsDataCruncher.R
\name{flowDataCruncher}
\alias{flowDataCruncher}
\title{This takes a matrix of data with certain columns and converts them into a 3 or 4 D matrix that can be read by Tim's pvheatmap package}
\usage{
flowDataCruncher(
  my_data,
  selected_type = NULL,
  type = NULL,
  group,
  subgroup = NULL,
  ID = NULL,
  ignore = NULL,
  combine_groups = NULL,
  control_group = NULL
)
}
\description{
This takes a matrix of data with certain columns and converts them into a 3 or 4 D matrix that can be read by Tim's pvheatmap package
}
