% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsDataLoader.R
\name{flowJoDataLoader}
\alias{flowJoDataLoader}
\title{This grabs flowjo files and extracts data
I could make this combine the files, but it'll be easier to use and more versitle if we just combine them outside this function}
\usage{
flowJoDataLoader(sample_list, file, fcs_files, sample_info)
}
\description{
This grabs flowjo files and extracts data
I could make this combine the files, but it'll be easier to use and more versitle if we just combine them outside this function
}
