% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsBasicHeatmap.R
\name{PVHeatmap}
\alias{PVHeatmap}
\title{This makes a heatmap with integrated significance values (one line for p<0.05 and two lines for p<0.005)}
\usage{
PVHeatmap(
  values,
  crunched_data = FALSE,
  pvalues = NULL,
  sem_high = NULL,
  sem_low = NULL,
  transpose = FALSE,
  title = NULL,
  font_size = NULL,
  font_size_x = NULL,
  font_size_y = NULL,
  font_size_title = NULL,
  nudge_title_verticle = 0,
  nudge_title_horizontal = 0,
  nudge_x_axis_verticle = 0,
  nudge_x_axis_horizontal = 0,
  nudge_y_axis_verticle = 0,
  nudge_y_axis_horizontal = 0,
  nudge_map_verticle = 0,
  nudge_map_horizontal = 0,
  gap = 0.1,
  box_size = NULL,
  border_width = 1,
  sub_border_intensity = 0.5,
  sub_border_width = 1,
  combine_pvalue = FALSE,
  justify_y = 0,
  sort_y = NULL
)
}
\description{
It's my first attempt at a package, don't expect to much from it! 2021-11-15 and has very little easy control for now, just as a utility for specific needs of the lab
}
\details{
Call the function timsBasicHeatmap(means, pvalues)

means is a matrix of means with appropriate column and row names that will be used in the heatmap graphic

pvalue is a matching matrix with the pvalue that accompanies the mean, The size and layout of the matrices must be identical

Columns should be phenotypes, rows should be groups
}
