% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsBasicHeatmap.R
\name{timsBasicHeatmap}
\alias{timsBasicHeatmap}
\title{This makes a heatmap with intergrated significance values (one line for p<0.05 and two lines for p<0.005)}
\usage{
timsBasicHeatmap(means, pvalues)
}
\description{
It's my first attempt at a package, don't expect to much from it! 2021-11-15 and has very little easy control for now, just as a utility for specific needs of the lab
}
\details{
Call the function timsBasicHeatmap(means, pvalues)

means is a matrix of means with appropriate column and row names that will be used in the heatmap graphic

pvalue is a matching matrix with the pvalue that accompanies the mean, The size and layout of the matrices must be identical

Columns should be phenotypes, rows should be groups
}
